- leaderboard fixes - denshokan fix

+++ packages/leaderboard/src/leaderboard.cairo
@@ -7,7 +7,7 @@ pub mod leaderboard {
     #[derive(Drop, Serde, Copy)]
     pub struct LeaderboardConfig {
         /// Maximum number of entries allowed
-        pub max_entries: u8,
+        pub max_entries: u32,
         /// Whether lower scores are better (true) or higher scores are better (false)
         pub ascending: bool,
         /// Whether to allow ties (same score, different entries)
